+++ stylesheets/library/index.ts
@@ -2,5 +2,7 @@ export * from "./albumDetail";
 export * from "./albums";
 export * from "./artists";
 export * from "./main";
+export * from "./playlistDetail";
+export * from "./playlists";
 export * from "./songs";